refactor: name project preview boundary explicitly
