--- packages/project-build/src/vision.ts
@@ -1,1 +1,0 @@
-export * from "./visual/preview";
